+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   `Link`: deprecated `rightIcon`/`leftIcon` props (use nested icons instead)
+-   `Link` docs: rework documentation page
+
 
 ## [3.12.0][] - 2025-03-19