+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.13.0][] - 2025-03-27
+
 ### Added
 
 -   `InputLabel`: add `typography` prop
@@ -2237,7 +2239,8 @@ _Failed released_
 [3.7.0]: https://github.com/lumapps/design-system/compare/v3.6.8...v3.7.0
 [3.6.8]: https://github.com/lumapps/design-system/compare/v3.6.7...v3.6.8
 [3.6.7]: https://github.com/lumapps/design-system/tree/v3.6.7
-[Unreleased]: https://github.com/lumapps/design-system/compare/v3.12.0...HEAD
+[Unreleased]: https://github.com/lumapps/design-system/compare/v3.13.0...HEAD
+[3.13.0]: https://github.com/lumapps/design-system/compare/v3.12.0...v3.13.0
 [3.12.0]: https://github.com/lumapps/design-system/compare/v3.11.3...v3.12.0
 [3.11.3]: https://github.com/lumapps/design-system/compare/v3.11.2...v3.11.3
 [3.11.2]: https://github.com/lumapps/design-system/compare/v3.11.1...v3.11.2